#14 fix followup 7: float32 cast for in-capture cumsum on (dp_size,)
Pod fr67w cycle showed the cumsum error reproduces from INSIDE captured
graphs — my followup-5 guard sends in-capture calls down the GPU
torch.cumsum(int64) path which gets baked into the cuda graph and then
fires the B200 SM_100 launch-config bug on replay (first inference).

Workaround: cast int64 -> float32 for cumsum, cast back. The fp32 cumsum
kernel uses a different launch config that works on B200 SM_100 for the
small (dp_size,) shape. float32 is lossless for values up to 2^24 which
covers any realistic global token count.

Also kept the existing eager-path CPU prefix sum (negligible cost at
dp_size=8) so non-captured launches stay safe.

4 occurrences patched (logits_processor.py: capture-path + None-mirror
fallback; dp_attention.py: same two sites in get_dp_local_info).

+++ python/sglang/srt/layers/logits_processor.py
@@ -221,11 +221,11 @@ class LogitsMetadata:
         # and re-materialize on device. dp_size <= 8 so the cost is negligible.
         _src = self.global_num_tokens_for_logprob_gpu
         if torch.cuda.is_current_stream_capturing():
-            cumtokens = torch.cumsum(_src, dim=0)
+            cumtokens = torch.cumsum(_src.to(torch.float32), dim=0).to(_src.dtype)
         else:
             _src_cpu_list = self.global_num_tokens_for_logprob_cpu
             if _src_cpu_list is None:
-                cumtokens = torch.cumsum(_src, dim=0)
+                cumtokens = torch.cumsum(_src.to(torch.float32), dim=0).to(_src.dtype)
             else:
                 _acc = 0
                 _prefix = []
